pyproject.toml
[project]
authors = [{name = "WhirlwindOfLight", email = "144183077+WhirlwindOfLight@users.noreply.github.com"}]
dependencies = ["python-uinput>=1.0.1,<2", "pyudev>=0.24.4,<0.25"]
name = "VM-Controller"
requires-python = ">= 3.11"
version = "0.1.0"

[project.scripts]
vm-controller = "vm_controller.main:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
vm_controller = { path = ".", editable = true }

[tool.pixi.tasks]
